--- src/InfoComponents/InfoData.json
@@ -1,15 +1,0 @@
-[
-    {
-        "title1": "$1.13 million",
-        "content1": "Americans spend $1.13 million on unused gym memberships annually.",
-        "title2": "$1.9 billion",
-        "content2": "Australians are wasting $1.9 billion per year on unused gym memberships",
-        "title3": "£4 billion",
-        "content3": "Britons waste more than £4 billion a year on unused gym memberships"
-    },
-    {
-        "title": "So much money is being tossed away for nothing. We want to",
-        "yellowPart": "change it.",
-        "buttonText": "Read more about us"
-    }
-]
